# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Memtara API
  version: 0.1.0
  summary: Zero-knowledge attribute disclosure, and the signed attestations that carry its verdicts off-platform.
  description: |
    Memtara lets a consumer prove a *predicate* about themselves — income above a
    threshold, residency valid, no sanctions match — without disclosing the
    attribute behind it. Proofs are generated on the holder's device against one
    of four Noir circuits; this API authenticates users, syncs the encrypted
    vault, orchestrates the disclosure-request lifecycle, verifies submitted
    proofs against the real compiled circuits via Barretenberg, and issues
    short-lived Ed25519-signed attestations that any JOSE-speaking relying party
    can validate offline.

    **This API never generates proofs.** Generating one requires the plaintext
    attribute as a witness, so a server that could prove would be a server that
    holds plaintext. See `ARCHITECTURE.md`. Every "verified" verdict here comes
    from running the real circuit's verifier over bytes a client produced.

    **Compliance context.** `docs/REGULATORY_MATRIX.md` maps this system, clause
    by clause, against the CBUAE *Guidance Note on Consumer Protection and the
    Responsible Adoption and Use of AI and ML by LFIs* (`CBUAE_EN_6958_VER1`).
    Predicate responses carry the clause ids each disclosure is evidence for.
  license:
    name: See repository
  contact:
    name: Memtara
...
servers:
...
security:
  - sessionToken: []
paths:
...
  /api/v1/issue-proof:
    post:
      tags: [Proof issuance]
      summary: Issue a signed attestation that a predicate was proven
      description: |
        Two request shapes, both real circuit evaluations:

        * **Fresh** — supply `disclosure_request_id`, `public_inputs` and
          `proof`. The proof is checked against the circuit's verification key
          via `bb verify` right now, the request's nonce is consumed, and a
          token is issued only on a genuine cryptographic pass.
        * **Attested** — omit `public_inputs` and `proof`. The most recent proof
          that already passed verification for this `(user, circuit)` pair is
          re-attested. This exists because a proof is single-use by design (its
          nonce is burned), while a bank may legitimately need a second token
          for the same verified fact inside one onboarding session.

        Both paths write a `proof_token_issued` event, carrying the
        `regulatory_audit_id` and which path was taken, to the hash-chained
        audit log.

        **The issued token cannot be revoked inside its lifetime.** That is the
        deliberate cost of offline validation; see the §6(f) caveat in
        `docs/REGULATORY_MATRIX.md`. Revoke the disclosure request before
        issuance, or the user's session to prevent re-issuance.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueProofRequest"
            examples:
              fresh:
                summary: Fresh evaluation against a pending disclosure request
                value:
                  user_id: 8f14e45f-ceea-467a-9f2a-1c1b2f3d4e5a
                  predicate: income_gte_threshold
                  disclosure_request_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  public_inputs: ["0x66c1e0a0", "0x66c1e6e0", "0x1a2b…", "0x0f4240", "0x1e8480", "0x…", "0x…", "0x…"]
                  proof: bXktcHJvb2YtYnl0ZXM
              attested:
                summary: Re-attest a proof this user already had verified
                value:
                  user_id: 8f14e45f-ceea-467a-9f2a-1c1b2f3d4e5a
                  predicate: residency_valid
      responses:
        "200":
          description: Attestation issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueProofResponse"
        "400":
          description: |
            Unknown predicate; `public_inputs`/`proof` supplied without each
            other; wrong public-input arity; nonce mismatch; the disclosure
            request is for a different circuit than the predicate requires; or
            the proof failed cryptographic verification (no token is minted, and
            the failure is itself recorded).
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          description: No such disclosure request.
          $ref: "#/components/responses/Error"
        "409":
          description: |
            The disclosure request is no longer pending; the nonce has already
            been consumed; or (attested path) no verified proof exists for this
            user and circuit.
          $ref: "#/components/responses/Error"
...
  /api/v1/submit-wealth-proof:
    post:
      tags: [Structured-product suitability]
      summary: Answer a suitability assessment with a proof
      description: |
        The client's half. Authorised by either the org that opened the request
        or the user it is for.

        Checks, in order, before anything is signed:

        1. every submitted public input matches the terms the org registered;
        2. `current_time` falls inside the server's assessment window;
        3. `vault_root` equals the root this user actually synced via
           `PUT /vault` — without which the circuit's Merkle limb proves only
           that the figures are consistent with *some* tree, and a client can
           build one to order;
        4. `bb verify` against the `wealth_suitability` verification key, with
           the nonce consumed atomically on a pass;
        5. **public input 11 is read.**

        Step 5 is not optional. `bb verify` answers "was this proof correctly
        constructed", not "is the client suitable" — a proof that the client
        *failed* verifies exactly as cleanly as one that they passed. A token
        is issued either way, carrying the real verdict in `suitable`; a
        declined recommendation needs its evidence as much as an approved one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitWealthProofRequest"
      responses:
        "200":
          description: |
            Assessment complete. Check `suitable` — a `200` means the proof
            verified, not that the client qualified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitWealthProofResponse"
        "400":
          description: |
            Public inputs disagree with the registered terms; `current_time`
            outside the window; `vault_root` is not this user's committed root;
            wrong arity; nonce mismatch; the proof failed verification; or the
            circuit output was neither 0 nor 1.
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          description: No such suitability request.
          $ref: "#/components/responses/Error"
        "409":
          description: |
            Already assessed; the nonce has been consumed; or the user has no
            synced vault to check the assessment against.
          $ref: "#/components/responses/Error"
...
  /orgs/{id}/audit-log:
    get:
      tags: [Organizations]
      summary: This org's slice of the hash-chained audit log
      description: |
        Rows whose `ref_id` names one of this org's disclosure requests. The
        chain itself is global (it commits to the true interleaving across all
        orgs, which is strictly stronger than per-org chains); this endpoint
        filters the view, not the chain.
      security: [{ orgApiKey: [] }]
      parameters: [{ $ref: "#/components/parameters/OrgId" }]
      responses:
        "200":
          description: Audit entries in chain order.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditEntry" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: |
        An opaque, revocable session token from any auth flow. Opaque rather
        than a JWT precisely so it can be killed in one query — the product's
        revocation guarantees depend on it.
    orgApiKey:
      type: http
      scheme: bearer
      description: |
        A relying party's API key, presented in the same `Authorization: Bearer`
        header as a session token. Routes readable by "the org that created it
        or the user it is for" try the presented credential as both.
  parameters:
...
    OrgId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Error.
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
  schemas:
...
    IssueProofRequest:
      type: object
      required: [user_id, predicate]
      properties:
        user_id: { type: string, format: uuid }
        predicate:
          type: string
          description: Must be one of `GET /api/v1/predicates`.
        disclosure_request_id:
          type: string
          format: uuid
          description: Required on the fresh path; optional on the attested path, where it narrows which prior proof to re-attest.
        public_inputs:
          type: array
          items: { type: string, pattern: "^0[xX][0-9a-fA-F]{1,64}$" }
          description: One `0x`-prefixed hex field element per public input, in the circuit's declared order. Supply with `proof` or omit both.
        proof:
          type: string
          description: base64url-no-pad `bb` proof bytes.
    IssueProofResponse:
      type: object
      required: [proof_token, expires_in, regulatory_audit_id]
      properties:
        proof_token:
          type: string
          description: Compact JWS (EdDSA). Claims are described by `ProofTokenClaims`.
        expires_in:
          type: integer
          description: Seconds. 300 by default.
          example: 300
        regulatory_audit_id:
          type: string
          format: uuid
          description: Echoed from the token so a caller can log it without decoding the JWT.
...
    SubmitWealthProofRequest:
      type: object
      required: [request_id, public_inputs, proof]
      properties:
        request_id: { type: string, format: uuid }
        public_inputs:
          type: array
          minItems: 12
          maxItems: 12
          items: { type: string, pattern: "^0[xX][0-9a-fA-F]{1,64}$" }
        proof:
          type: string
          description: base64url-no-pad `bb` proof bytes (14,656 bytes for this circuit).
    SubmitWealthProofResponse:
      type: object
      required: [proof_token, expires_in, suitable, product_isin, regulatory_audit_id]
      properties:
        proof_token:
          type: string
          description: Compact JWS (EdDSA), carrying `suitable`, `product_isin` and `dfsa_rules` in addition to the usual claims.
        expires_in: { type: integer, example: 300 }
        suitable:
          type: boolean
          description: |
            The circuit's verdict, echoed outside the token so a caller can act
            on it without decoding the JWT. The authoritative copy is the
            signed `suitable` claim inside the token; this one is convenience,
            not evidence.
        product_isin: { type: string }
        regulatory_audit_id: { type: string, format: uuid }
...
    AuditEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        event_type:
          type: string
          examples: [disclosure_request_created, proof_verified, proof_verification_failed, proof_token_issued]
        ref_id: { type: string, format: uuid, nullable: true }
        event_hash: { type: string, description: base64url-no-pad SHA-256. }
        prev_hash: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...